refactor: 새로운 파일로 코드 분리
