+++ CheongHa/CheongHa/CleanArchitecture/Presentation/Community/CommunityViewController.swift
@@ -11,101 +11,6 @@ import FlexLayout
 import RxSwift
 import RxCocoa
 
-enum CommunityCategory: Int, CaseIterable {
-    case free = 0
-    case info
-    case question
-    case life
-    case hobby
-    case economy
-    
-    var iconNotSelectedImage: ImageResource {
-        switch self {
-        case .free:
-            return .icFreeTagNotSelected
-        case .info:
-            return .icInfoTagNotSelected
-        case .question:
-            return .icQuestionTagNotSelected
-        case .life:
-            return .icLifeTagNotSelected
-        case .hobby:
-            return .icHobbyTagNotSelected
-        case .economy:
-            return .icEconomyTagNotSelected
-        }
-    }
-    
-    var iconSelectedImage: ImageResource {
-        switch self {
-        case .free:
-            return .icFreeTagSelected
-        case .info:
-            return .icInfoTagSelected
-        case .question:
-            return .icQuestionTagSelected
-        case .life:
-            return .icLifeTagSelected
-        case .hobby:
-            return .icHobbyTagSelected
-        case .economy:
-            return .icEconomyTagSelected
-        }
-    }
-    
-    var categoryName: String {
-        switch self {
-        case .free:
-            return "자유"
-        case .info:
-            return "정보"
-        case .question:
-            return "질문"
-        case .life:
-            return "생활"
-        case .hobby:
-            return "취미"
-        case .economy:
-            return "경제"
-        }
-    }
-}
-
-final class CommunityTabButton: UIView {
-    
-    let categoryIconView = UIImageView()
-    
-    let categoryLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = .gray2
-        label.font = .systemFont(ofSize: 10)
-        return label
-    }()
-    
-    init(category: CommunityCategory) {
-        super.init(frame: .zero)
-        
-        self.flex.define { flex in
-            flex.addItem(categoryIconView).size(28).margin(6)
-            flex.addItem(categoryLabel).marginHorizontal(11)
-        }
-        .direction(.column).justifyContent(.center).alignItems(.center)
-        
-        categoryIconView.image = UIImage(resource: category.iconNotSelectedImage)
-        categoryLabel.text = category.categoryName
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        
-        self.flex.layout()
-    }
-}
-
 final class CommunityViewController: UIViewController {
     
     private let baseContainer = UIView()
